smooth interpolation for saber length, fixed overrided shader which did not apply to speed trail, fixed laggy speed trail on low frame update

+++ codemp/qcommon/q_shared.h
@@ -876,7 +876,7 @@ typedef struct
 	saberTrail_t	trail;
 	int			hitWallDebounceTime;
 	int			storageTime;
-	float		extendDebounce;
+	double		extendDebounce;
 } bladeInfo_t;
 #define MAX_BLADES 8
